Drop dead navigate/executeAction from HarmonicClient; extract frontmatter parsers
After the agent-runner migrated all agent-acting calls to /mcp via
McpClient, HarmonicClient.navigate and executeAction had no production
callers — the only test usage was a stub raising "deprecated" errors.
Remove them.

HarmonicClient now exposes only fetchChatHistory, which the runner uses
to build the LLM prompt at chat-turn start and stays on the HMAC
/internal/... path because it isn't a tool the agent invokes.

Move parseAvailableActions and parseResolvedPath into a new
core/MarkdownFrontmatter module so they live alongside other markdown
parsers and don't depend on a service module they share no semantics
with. The functions are pure data parsers; McpClient imports them
where it consumes fetch_page responses.

While here:
- Comment on the asymmetric error handling in McpClient (fetchPage
  throws on tool error, executeAction returns success: false — both are
  correct; the reason is that fetchPage's downstream consumers depend on
  parsed-frontmatter fields that don't exist on an error response).
- Add unit tests for HarmonicClient.fetchChatHistory, which had no
  direct coverage before (only indirect via AgentLoop mocks).
